+++ hw16.09/ex7.py
@@ -1,5 +1,5 @@
-x = int(input())
-y = int(input())
+x = float(input())
+y = float(input())
 sum = 0
 for i in range(x, y+1):
     if i % 5 == 0:
